src/backend/requirements_monitoring.txt
# OpenTelemetry core dependencies
opentelemetry-api
opentelemetry-sdk

# OpenTelemetry instrumentations
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-celery

# OpenTelemetry exporters
opentelemetry-exporter-otlp

# FastAPI-specific dependencies
pydantic-settings 